requirements.txt
openai
python-dotenv
backoff
transformers
torch
sentencepiece